+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 # 1.5
 ```
-    * Fixed the URL  to the current plugin repo, when check for new version.
+    * Fixed the URL to the current plugin repo, when check for new version.
     * Changed Sandbox REST API URL.
     * Added option to change SimplyConnect theme.
+    * Added option to change APM window type.
 ```
 
 # 1.4